chore: refactor row pinning feature
